Make building with Docker easier (#407)
* Make building with Docker easier
* More Docker local build improvements

+++ docker-build/scripts/01_clone.sh
@@ -1,37 +1,40 @@
 set -e
 
-cd /
-UBUNTU_VER=18.04
+if [ ! ${LOCAL_BUILD} ]; then
+    git config --global user.name  "Docker SpringRTS Build"
+    git config --global user.email "dockerbuild@beyondallreason.info"
 
-# Do not use git depth parameter cause git describe later will not work as expected
+    rm -rf "${SPRING_DIR}"
 
-git config --global user.name  "Docker SpringRTS Build"
-git config --global user.email "dockerbuild@beyondallreason.info"
+    echo "---------------------------------"
+    echo "Cloning SpringRTS from: ${SPRINGRTS_GIT_URL}"
+    echo "Using branch: ${BRANCH_NAME}"
+    echo "---------------------------------"
 
-rm -rf "${SPRING_DIR}"
+    # Do not use git depth parameter cause git describe later will not work as expected
+    CMD="git clone "${SPRINGRTS_GIT_URL}" "${SPRING_DIR}""
 
-echo "---------------------------------"
-echo "Cloning SpringRTS from: ${SPRINGRTS_GIT_URL}"
-echo "Using branch: ${BRANCH_NAME}"
-echo "---------------------------------"
+    echo "Command: ${CMD}"
 
-CMD="git clone "${SPRINGRTS_GIT_URL}" "${SPRING_DIR}""
+    ${CMD}
 
-echo "Command: ${CMD}"
-
-${CMD}
-
-cd "${SPRING_DIR}"
-git checkout "${BRANCH_NAME}"
-git submodule update --init --recursive
-
-if [ "${PLATFORM}" == "windows-64" ]; then
-    git clone "${SPRINGRTS_AUX_URL_PREFIX}/mingwlibs64.git" mingwlibs64
-elif [ "${PLATFORM}" == "linux-64" ]; then
-    git clone "${SPRINGRTS_AUX_URL_PREFIX}/spring-static-libs.git" -b $UBUNTU_VER spring-static-libs
-else
-    echo "Unsupported platform: '${PLATFORM}'"
-    exit 1
+    cd "${SPRING_DIR}"
+    git checkout "${BRANCH_NAME}"
+    git submodule update --init --recursive
 fi
 
-cd "${SPRING_DIR}"
+if [ "${PLATFORM}" == "windows-64" ]; then
+    LIBS_DIR="/mingwlibs64"
+    LIBS_BRANCH=master
+elif [ "${PLATFORM}" == "linux-64" ]; then
+    LIBS_DIR="/spring-static-libs"
+    LIBS_BRANCH=18.04
+fi
+
+if [ ! -d "${LIBS_DIR}/.git" ]; then
+    git clone "${SPRINGRTS_AUX_URL_PREFIX}${LIBS_DIR}.git" -b "${LIBS_BRANCH}" --depth 1 "$LIBS_DIR"
+else
+    cd "${LIBS_DIR}"
+    git pull
+    git checkout ${LIBS_BRANCH}
+fi
